requirements.txt
browser-use
fastapi
uvicorn
python-dotenv
pydantic
openai
neo4j
playwright
httpx
